requirements.txt
python-dotenv==0.13.0
Flask==1.0.2
Flask-API==1.0
Flask-Cors==3.0.8
Flask-Excel==0.0.7
Flask-Ext==0.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
jsonify==0.5
requests==2.24.0
pymongo==3.11.0
dnspython==2.0.0